Rename `zkevm` crate to `prover`.
