+++ .gitmodules
@@ -1,3 +1,3 @@
-[submodule "test-traces"]
-	path = zkevm/tests/traces
+[submodule "prover/tests/traces"]
+	path = prover/tests/traces
 	url = https://github.com/scroll-tech/test-traces
