+++ Makefile
@@ -24,7 +24,7 @@ test: ## Run tests for all the workspace members
 
 bridge-test:
 	cargo build --release
-	./target/release/prove --params=./test_params --seed=./test_seed --trace=zkevm/tests/traces/bridge --agg=true
+	./target/release/prove --params=./test_params --seed=./test_seed --trace=prover/tests/traces/bridge --agg=true
 
 test-super-trace: ## test super circuit with real trace
 	cargo test --features prove_verify --release test_prove_verify
