+++ bin/Cargo.toml
@@ -21,7 +21,7 @@ serde_derive = "1.0"
 serde_json = "1.0.66"
 tokio = { version = "1", features = ["full"] }
 types = { path = "../types" }
-zkevm = { path = "../zkevm" }
+prover = { path = "../prover" }
 
 [[bin]]
 name = "setup"
