+++ prover/src/capacity_checker.rs
