+++ prover/src/circuit.rs
