+++ prover/src/circuit/builder.rs
