+++ prover/src/circuit/state_circuit.rs
