+++ prover/src/circuit/zktrie_circuit.rs
