+++ prover/src/prover/outer_circuit.rs
