+++ prover/src/prover/outer_proof.rs
