+++ prover/src/prover/util.rs
