+++ prover/tests/extra_traces/hash_precompile_1.json
