+++ prover/tests/extra_traces/hash_precompile_2.json
